tests: add a unit test framework
This provides a mechanism to run in-tree unit tests by typing
'make check'

The unit test framework is rather simplistic, but it's just so that we
can register a bunch of tests and execute them.  Something more
sophsticated can be cooked up later (if needed).

A simple example is provided.

Signed-off-by: Aaron Conole <aconole@redhat.com>