lldpad: fix detection of other lldpad instance
kill() can return a non-zero value on failure, and when it does errno will
be set to a number of values.  Two of these values (EINVAL and EPERM)
indicate that the kill() call was unsuccessful.  However, ESRCH can also
be set - in that case it means that the process isn't found.

This change excepts ESRCH from being fatal to the kill() call.  This allows
lldpad to restart in certain cases.

Signed-off-by: Aaron Conole <aconole@redhat.com>