autotest: add new test suite infrastructure
Typically, testing is confined to individual developers manually starting, running
and validating changes.  This isn't very repeatable, and doesn't allow for auditing
what kinds of testing are done.  Automation is not well shared.  The state of
public testing for the project isn't great.

As a first step to resolving this, introduce a mechanism to run test suites defined
in automake that can setup network namespaces and run various tests, using utilities
like netcat / tcpdump to send/receive data and interact with the openlldp suite of
tools.

Signed-off-by: Aaron Conole <aconole@redhat.com>